Add `ExternalChatSendEvent` class
A new class called `ExternalChatSendEvent` has been created in the package de.ayont.lpc.event. This class is an event handler which extends the `Event` class and implements the `Cancellable` interface. It contains methods related to chat messages in the application such as setting message format, retrieving the unique identifier, and handling event cancellation.